Prevent default action when clicking on a day.
This avoids a problem when the date picker is located inside an
HTML <label> element. The default action of clicking on a
(block-level) label is to focus an input inside the label
(either the one explicitly indicated by the label's "for"
attribute or the first one). The result is that the date picker
is blurred and then immediately refocused -- and therefore
stays open.